Several efficiency changes as well as additional utility functions for the soap server -- get projects, get networks etc. Added ProjectSummary complexmodel to aid in just displaying available projects without loading the whole things.
Added 2 tables to the DB schema for ownership -- tProjectOwner and tDatasetOwner.
This is essential for saying: Get projects for user 'X', and share project with user 'Y'.

Added some new tests, and fixed the tests to make them faster by reducing the number
of soap connections opened.

Finally added constraint handlers to the soap server.


git-svn-id: svn+ssh://ec2-54-229-95-247.eu-west-1.compute.amazonaws.com/home/svn/svnrepository/HYDRA@123 5b2e1b5e-a1f6-4386-8dd1-2f88c1042e8b